+++ test/unittest/util.jl
@@ -15,6 +15,8 @@ function test_firstarg()
   @test mx._firstarg(:(f(x::mx.NDArray{T}, y) where {T})) == :x
   @test mx._firstarg(:(f(x::mx.NDArray{T,N}, y) where {T,N})) == :x
   @test mx._firstarg(:(f(x::mx.NDArray{T,N} where {T,N}, y))) == :x
+  @test mx._firstarg(:(broadcast_(::typeof(asin), x::mx.NDArray))) == :x
+  @test mx._firstarg(:(broadcast_(::typeof(asin), x::mx.NDArray, y::mx.NDArray))) == :x
 end  # function test_firstarg